+++ sakura_core/charset/ShiftJis.h
@@ -44,7 +44,7 @@ public:
 // S_GetEolはCodeBaseに移動	2010/6/13 Uchi
 	// 2005-09-02 D.S.Koba
 	// 2007.08.14 kobake MemoryからShiftJisへ移動
-	static int GetSizeOfChar(const char* pData, int nDataLen, int nIdx); // 指定した位置の文字が何バイト文字かを返す
+	static size_t GetSizeOfChar(const char* pData, size_t nDataLen, size_t nIdx); // 指定した位置の文字が何バイト文字かを返す
 
 protected:
 	// 実装
